Add a custom parser for AUXV entries for CHERI binaries.
In particular, AUXV entries under CHERI are a long followed by
padding and a capability.  For AUXV entries that hold pointers,
extract the virtual address and return that.  This fixes 'info auxv'
to generally be correct and also permits gdb to find the initial
location of rtld via AT_BASE.