Add a hack to ignore DW_AT_bit_size for CHERI pointers.
This appears to be a regression in the recent compiler upgrade as clang
is now generating DW_AT_bit_size attributes for pointers which gdb
interprets to mean that they are bitfields.  OTOH, it seems that clang
is now emitting DW_AT_byte_size for pointers which might mean that the
previous hack for CHERI can be removed.