Fix is_cheri() for live MIPS-only processes on a CHERI CPU.
The is_cheri() function only checked to see if the target supplied a
cap0 register ($C0) which is true if the host supports PTRACE_CAPREGS or
if an executable has the CHERI ELF flag set.  Most of the 'is_cheri()'
checks in mips-tdep.c are probably harmless for a MIPS binary as they
are related to extra instructions.  However, some of the places (using
$PCC instead of $PC and modifying $SP and $S8 to add the cursor of $C11)
are not correct for MIPS binaries.

To fix, strengthen the is_cheri() check to require that the current binary
has the CHERI ELF flag set.  This won't quite do the right thing for
attaching to a CHERI ABI process without a valid executable but should be
correct for any target when the executable is available.