Improve handling of capability registers in cores.
In particular, use a target description with capability registers if the
core dump note exists instead of depending on the CheriABI ELF flags.
This means that cap regs can now be examined for hybrid MIPS64 binaries
(and possibly any MIPS64 binary running on a CHERI CPU if the kernel
includes the relevant ELF core note).  The size of the cap registers in
this case is inferred from the note's size rather than flags in the ELF
header.