First cut at handling new threads while stepping.
Remember the ptid passed to resume to determine if the current inferior is
being single stepped when events are reported from wait().

When a new thread reports an event while single stepping a different thread,
suspend the new thread via PT_SUSPEND and explicitly continue the process
to wait for an event from the stepping thread.

Checkpoint more WIP on the non LWP events case.

- Add an unordered_map<> mapping ptid_t objects to a struct target_waitstatus
  to cache the event for suspended threads.
- Don't ever PT_RESUME a suspended thread.
- If a matching suspended thread is eligible at resume time, don't
  PT_CONTINUE / PT_STEP the process but instead let fbsd_wait report the
  event from the suspended thread directly.
- If fbsd_wait needs to continue the process because it is ignoring a
  reported event, then invoke fbsd_resume in the non-LWP-events case.