Use a register's name as the key instead of the struct tdesc_reg pointer.
tdesc_use_registers uses a temporary hash table to determine which
registers from the desired target description are already present in
the architecture's register set.  It then assigns register numbers to
any new registers from the target description.

Previously, the hash table used the value of the struct tdesc_reg
pointer to determine if a register from the target description matched
an existing register in the architecture's register set.  However, if
the same register was in both sets with different register numbers
each set would use a different register structure.  To handle this
case, use the register name as the key to identify registers instead
of the struct tdesc_reg pointer.