Handle typedefs in the CHERI pointer bitsize hack.
The hack to ignore DWARF bitsize attributes on CHERI pointers only worked
for types that were a pointer, but not for typedef'd types such as
caddr_t.  Use 'check_typedef()' to resolve the type through any typedef's
to the root type to determine if it is really a pointer.  While here,
use the TYPE_CODE() helper macro as well.