Be more cautious about setting CheriABI handlers.
Only set CheriABI gdbarch methods if the current gdbarch is using a
target description that includes capability registers.  When attaching to
an existing process, GDB starts with a plain MIPS architecture and only
switches to a CHERI target description part way through at which point
it will create a second gdbarch.  Only install CheriABI-specific
gdbarch methods for this second gdbarch.  This fixes crashes when
trying to attach to a running process.