Improve fallback handling for the read of pcb_size.
If for some reason kgdb has only the kernel's debug info, a read of
pcb_size will simply return 0, resulting in an exception when unwinding
the stack of a thread that was on a CPU at the time the dump was
created.  In this case, fall back to using sizeof(struct pcb).