Rework kgdb -w support to only open /dev/mem writable.
Don't depend on the gdb global 'write_files' option which tries to
rewrite the kernel binary on close which can crash.  Instead, add an
optional '-w' flag to 'target vmcore' which opens the vmcore with
O_RDWR instead of O_RDONLY.  Change the kgdb '-w' option to set this
flag in the nested 'target vmcore' command rather than passing the
global '-w' option to gdb.