Properly skip printing attributes for NULL capabilities.
This broke when un-sealed capabilities grew a non-zero otype.  Since
there isn't a common way in the existing decompress_cap bits to determine
if attributes are valid, just add a separate bool variable that gets set
to false if the attributes are all-zero and only print attributes if
this bool is true.