[PPC] Fix ppcfbsd_trapframe_sniffer
It seems previous ppcfbsd_trapframe_sniffer() implementation assumed
it should work as FreeBSD's DDB, that compares the frame's return
address with some known sentinels.
In kgdb, however, the 'pc' obtained from 'this_frame' points to the
entry point of the function being executed in the frame.

Therefore, this patch checks if this_frame's function matches any of the
functions that may call C trap handling code.
This allows kgdb and remote kgdb to debug through traps.

This was tested on PowerPC64/AIM, but should work on PowerPC32/AIM
as well, given the trap code similarity between them. On BOOK-E,
"trap_common" function should probably be added to the list too,
but it would be better to test it first.