fbsd-nat: Implement async target support.
This is a fairly simple version of async target support.

Synchronous mode still uses blocking waitpid() calls in
inf_ptrace::wait() unlike the Linux native target which always uses
WNOHANG and uses sigsuspend() for synchronous operation.

Asynchronous mode registers an event pipe with the core as a file
handle and writes to the pipe when SIGCHLD is raised.  WNOHANG is
passed to waitpid when TARGET_WNOHANG is passed to the
inf_ptrace::wait.