regcache: Add more regcache_map helper methods.
A new variant of regcache_map_entry_size accepts a gdbarch parameter to
support register maps using variable sized registers where an entry size
of 0 means to use register_size (gdbarch, regnum) as the size instead.

regcache_map_offset returns the offset of a specified register in a
register map.  If the register not found, -1 is returned instead.