Add a custom osabi sniffer for FreeBSD kernels.
FreeBSD kernels are dynamic ELF executables (dynamic forces generation
of a dynamic section which is used by the boot loader and in-kernel
linker to resolve symbols used by kernel modules).  Kernels are not
executed by an interpreter as with typical dynamic ELF executables,
but an .interp section is still present.  FreeBSD's build system uses
an interpreter path of "/red/herring".  The sniffer uses this unique
quirk to identify kernels distinctly from userspace binaries.