Add shared library ops for FreeBSD kernel modules.
Walk the linker_files TAILQ in the main kernel to enumerate the loaded
modules.  This list functions much like the link_map linked-list in
runtime loaders.

Add a hook to initialize architecture-independent settings for FreeBSD
kernel architectures.  Currently it just registers the kernel module
shared library ops with the architecture.