gdbarch: Add hook to supply regcache from FreeBSD kernel pcb.
The 'struct pcb' (really a thread control block rather than a process
control block, but the old struct name is still used) contains saved
register state for suspended kernel threads.  The hook accepts the
address of an in-memory pcb and copies register values into the
supplied regcache.

This is used to populate initial register state for kernel threads
similar to fetching the register state of a stopped user thread via
ptrace.