fbsd-nat: Defer resume of inferiors.
For at least some tests in the testsuite, the core can resume multiple
individual threads belonging to the same process before waiting.  To
support this, change the resume method to record resumption requests
and defer resuming processes until either the commit_resumed target
method is called or until wait is called.

This requires tracking a set of resumed LWPs in the per-inferior
data rather than a single ptid_t to describe the set of resumed LWPs.